@@ -1,24 +1,33 @@
-// swift-tools-version: 5.7
+// swift-tools-version: 5.10
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
 
 let package = Package(
     name: "Domain",
-    platforms: [.iOS(.v16)],
+    defaultLocalization: "en",
+    platforms: [.iOS(.v17), .macOS(.v14)],
     products: [
         .library(
             name: "Domain",
-            targets: ["Domain"])
+            targets: ["Domain"]
+        )
     ],
     dependencies: [
+        .package(path: "../Toolkit"),
+        .package(path: "../CleanArchitecture"),
     ],
     targets: [
         .target(
             name: "Domain",
-            dependencies: []),
+            dependencies: [
+                "Toolkit",
+                "CleanArchitecture"
+            ]
+        ),
         .testTarget(
             name: "DomainTests",
-            dependencies: ["Domain"])
+            dependencies: ["Domain"]
+        )
     ]
 )
