@@ -1,25 +1,35 @@
-// swift-tools-version: 5.7
+// swift-tools-version: 5.10
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
 
 let package = Package(
     name: "Presentation",
-    platforms: [.iOS(.v16)],
+    defaultLocalization: "en",
+    platforms: [.iOS(.v17), .macOS(.v14)],
     products: [
         .library(
             name: "Presentation",
-            targets: ["Presentation"])
+            targets: ["Presentation"]
+        ),
     ],
     dependencies: [
-        .package(path: "../Domain")
+        .package(path: "../Domain"),
+        .package(path: "../UIComponents"),
+        .package(path: "../DesignSystem"),
     ],
     targets: [
         .target(
             name: "Presentation",
-            dependencies: ["Domain"]),
+            dependencies: [
+                "Domain",
+                "UIComponents",
+                "DesignSystem"
+            ]
+        ),
         .testTarget(
             name: "PresentationTests",
-            dependencies: ["Presentation"])
+            dependencies: ["Presentation"]
+        ),
     ]
 )
