@@ -1,5 +1,6 @@
-import XCTest
+import Testing
 @testable import Presentation
 
-final class PresentationTests: XCTestCase {
+@Test func example() async throws {
+    // Write your test here and use APIs like `#expect(...)` to check expected conditions.
 }
